added media tweets filter and likes tweets in user's profile, use nib file for TweetCell for better reusing, fixed the issue of favor and retweet api not working on actual device due to Uint overflow for tweet ids, fixed the infinite scrolling issue on the home timeline

+++ Twitter/Tweet.swift
@@ -10,8 +10,8 @@ import Foundation
 
 class Tweet{
     let tweetDict: [String: Any]!
-    var id: UInt{
-        return tweetDict["id"] as! UInt
+    var id: Int64{
+        return tweetDict["id"] as! Int64
     }
     
     var text: String{
